Refactor textMessageHandler tests to improve command routing and add new command handlers
- Updated command constants to include new commands for simulation and scraping.
- Mocked new command handlers for better separation of concerns.
- Reorganized test cases to route commands to their respective handlers.
- Added tests for new commands: /load_simulation, /trigger_scraping, and /get_current_simulation.
- Enhanced number message handling with additional test cases for edge cases.
- Removed redundant mocks and streamlined setup for clarity and maintainability.
